WT-2565: item 3573 on page at [write-check] is a corrupted cell
Cap the amount of data supplied to the raw compression engine at
(WT_BTREE>maxmempage /2), additionally, flag to the compression engine
that no more rows will be accumulated if the data has reached that size.

Edge cases include when there are no more rows to be accumulated, (but
that shouldn't happen, maxmempage is at least 50 x the maxleaf page).

Also, handle an EAGAIN return from the underlying compression engine
when there are additional rows to accumulate but the data size has grown
too large.