WT-2573: free of stack-allocated WT_REF
The error path in __wt_split_rewrite can free a WT_REF that was allocated
on the stack. Allocate the WT_REF instead, rewriting pages in memory isn't
a performance path, and it's a simple solution.

This change means we have never exercised this error path before. This is
an in-memory run which we've recently added to normal stress testing, but
I'm concerned about this, it seems unlikely a page-rewrite would fail in
normal testing.

Also fix a bug found by inspection, we weren't freeing allocated WT_REF
structures that didn't also have associated page images, which could lead
to memory leaks. There are failure cases where we would have allocated a
WT_REF and either not yet attached a page image or, in some cases, there's
no page image to attach. I'm calling the underlying __wt_free_ref() function
to free the WT_REF structure even when there's no associated page: that
looks safe, but regardless, I want to fix that function if this change fails,
not expicitly freeing the WT_REF structure in the split code.