WT-2577: core dump discarding non-existent addresses
We were using use the multiblock disk image to flag the existence of the
saved-update list, which isn't correct. In __split_multi_inmem(), we
clear the disk image reference when transferring ownership of the page
image to newly created split page, and when we reconcile that page again
(presumably because the split failed), we'll drop core trying to free
non-existent disk blocks. Use the saved-update list instead of the disk
image, we don't discard it during the split.