WT-2579: in-memory configurations break debugging support
In-memory configurations break the debugging support which wants to
write (potentially very large) chunks of information to files, for
example, dumping disk images, or in-memory pages, or the cache, and
writing that information into memory is useless.

Quit trying to make this work, revert the debugging functions to use
stdio calls directly.