WT-2581: assert multi->disk_image == NULL
The multiblock disk image may be NULL (it's handed off to another page
as part of in-memory page rewrites), the saved-update pointer may be
NULL (there may not have been any saved updates). Switch to using the
address field to determine if there are backing blocks that need to be
discarded, and always free any disk image or block of saved updates.