WT-2586: examples/ex_config is broken
The ex_config example code opens a "config:" URI and then does a
WT_CURSOR.next, but "config:" URIs don't support WT_CURSOR.next.

In 2010, the file was changed to create "table:access", but the cursor
URI wasn't switched to use it, it continued using "config:". The code
probably never worked, but when we changed the standard not-supported
cursor function to output an error message in WT-2349, the failure
became visible.