WT-2552 Minor cleanup of new pluggable file system API
* WT-2552 Add public API for pluggable filesystems

When swapping handling functions in the POSIX WT_FILE_HANDLE.fallocate
function, add a write-barrier so we never see an inconsistent set of
functions. It probably wouldn't matter, but let's not find out.

* WT-2552 Add public API for pluggable filesystems

Fix a couple of places in the in-memory code where we were redirecting
through the iface structure unnecessarily.

* WT-2552 Add public API for pluggable filesystems

Consistently do diagnostic asserts on read-only in the WiredTiger stub
functions, not in the underlying file-system/file-handle layer.