WT-2268 WT-2597 JSON load/dump Unicode fixes (#2749)
* WT-2268 Change dump representation so every byte represents a single 0x00-0xff
Unicode character. Otherwise arbitrary binary data (that does not conform to
Unicode) cannot be represented. Fix an off-by-one error in counting bytes
on input. Added some better Unicode tests.

* JSON Dump now uses a 'dump version' stamp, as well as showing WT version.
Dump input that is too old, and input that is too new will be rejected.

* In setting a key for dump, converting a JSON string must always occur first.

* When loading JSON, treat binary data just like strings.
Fixed an error in determining the string length of JSON in the presence
of Unicode strings.

* Add LSM tests to JSON dump/load testing.

* Add more extensive testing for dump/reload of JSON binary data.

* Byte arrays differ from strings for JSON input, they do not null terminate.
Handle escapes like '\n', '\t', etc. on input.

* Added tests of JSON dump/load of all byte codes.

* whitespace