WT-2673 Stop automatically increasing memory page max (#2760)
We used to ensure it was at least 50 times the leaf page size, enhancements to how we split pages mean that's no longer necessary.

While reviewing the code adjust the maximum page size allowed to ten percent of the cache size. It is more reasonable for workloads with multiple tables, and should only be relevant when applications configure tiny caches.