SERVER-23611 Improve signature of prepareExecution()
Instead of prepareExecution() taking two out-parameters and mutating
one of its in-parameters, it now explicitly takes ownership of the
CanonicalQuery passed in and returns a StatusWith<> that wraps the
three return values.  Forcing callers to std::move() their
CanonicalQuery and extract it back out should help make it more
obvious that it may be mutated.