SERVER-23830 Remove hardcoded pidfile path in init script header
This is a small change, the only affect of which is that systemd's
sysvinit compatibility implementation stops assuming this value for
the pidfile on rhel 7 and detects the pid automatically. For SUSE,
there's a separate init.d script. For non-systemd systems, the init
script parses mongod.conf for the pidfile.

(cherry picked from commit b68339f58d1168569ac46e8749429dd26e73748f)