WT-2729 Focus eviction on the largest trees in cache. (#2832)
Randomize visits to trees that use a tiny fraction of the cache.

Other general eviction optimizations.

Now that we are queuing more entries (potentially), make sure enough of
them become candidates.  Previously, a skewed distribution of read
generations could mean that only 10% of queue entries were considered.

Improve the efficiency of sorting the queue by calculating the score
once when pages are added to the queue.

Add a workload to exercise differential eviction from trees.

In order to do that, add range partitioning to wtperf.

Don't override icount in wtperf workloads with random_range set.