WT-2729 Focus eviction on the largest trees in cache (mongodb-3.2). (#2833)
Randomize visits to trees that use a tiny fraction of the cache.

Eviction optimizations.

Now that we are queuing more entries (potentially), make sure enough of
them become candidates.  Previously, a skewed distribution of read
generations could mean that only 10% of queue entries were considered.

Improve the efficiency of sorting the queue by calculating the score
once when pages are added to the queue.

Take care to bound the maximum eviction slot.