SERVER-4786 Allow specifying sample rate of slow queries
Ideally, we'd like to be able to set a lower slowms threshold so that we
can get a broader sample of queries being run against our
databases. However, dropping the threshold low enough to get useful data
results in an unacceptable amount of load.

This allows us to drop the threshold, but only capture a fraction of
queries.