WT-2730 Btree can incorrectly match key slots on new pages (#2848)
This problem can occur for both row and column store.

The WT_CURSOR_BTREE.rip_saved field potentially has the same problem
as the cip_saved field, initializing it on point-searches is wrong,
it should be initialized as a cursor moves to a new page.

* Clear cip_saved and rip_saved when starting to iterate from a search
position.  This wasn't necessary before because we cleared them in
__cursor_pos_clear(), but I removed that code.

In summary, we now clear them in the iteration code, both when starting
an iteration and when switching to a new page. That's correct because
they have nothing to do with searches so the clear doesn't belong in
__cursor_pos_clear(), and we have to do the clear when switching to a
new page regardless, __cursor_pos_clear() isn't called when switching
to a new page.

(cherry picked from commit 1b6a9220c3ce948f902c6bc44660b76982c7e621)