WT-2737 Page scrubbing: more fixes. (#2899)
* WT-2737 Page scrubbing: more fixes.

* Swap endianness of the page header back after writing (for zSeries).
* Deal with restoring updates when the WT_EVICT_SCRUB flag is not set:
  we want to save disk images for only those pages that have updates to
  restore.  Unfortunately, that isn't easy to figure out down in the
  guts of raw compression.

* typo

* Move __wt_page_header_byteswap() calls to immediately before/after the write
call, so error paths don't leave the caller's content in an unexpected format.

* For now, WT_EVICT_UPDATE_RESTORE implies WT_EVICT_SCRUB; we may be able to
relax this in the future.

* whitespace

* If raw compression gives up, the split boundary's already-compressed
flag won't be true, but we will have a saved disk image. We can change
raw compression to only create disk images if the already-compressed
flag will be true, or always create a disk image in raw-compression,
regardless of whether or not we actually compressed data.  I'm going
with the latter, it's simpler.

* Fix a comment.