WT-2737 page scrubbing: more fixes (#2901)
Rework 8bfe54b: we have to swap the page-header before calculating the
checksum and I don't want to leave the caller's data changed in the case
of an error. I also don't want to rework all the error handling in the
__wt_block_write_off() function, so wrap the block-write functionality
in a new function that only handles page-header swapping.