WT-2783 Clean up wtperf configuration object management (#2957)
The multi-database wtperf feature implementation assigns memory from the
base CONFIG structure to the CONFIG copies, but then individually frees
the memory referenced by the copies, resulting in OS X complaints that
memory is freed more than once.

Rename the config_assign() function to be config_init(), it does more
than simply assign the previous values.

In config_init(), create copies of the configuration strings in each new
CONFIG structure. There are also a few places where CONFIG.home and
CONFIG.monitor_dir are set (for example, to the value of the command
line -h and -m options), change that code to free/allocate so we maintain
a local copy of the strings in every CONFIG structure.

Change config_init() to copy the CONFIG.home and CONFIG.monitor_dir
fields along with the rest of the CONFIG structure, they get freed in
config_free(), so they should be copied in config_init().