WT-2336 Add a test validating schema operations via file system call monitoring (#3059)
Take advantage of the above to create a maintainable .run file that works
on systems without ftruncate. Use macros to avoid using so many #ifdefs.

When output differences are detected, report the output's line number (or
range of lines numbers we tried to match) to make it easier to diagnose
differences.

Add O_NOATIME to list of predefines to check.  The probe needs to be
compiled -D_GNU_SOURCE on Linux.  Use all probed variables in the symbol
table when looking at strace/dtruss output, as strace uses symbolic flag names.