WT-2402 Pad structures to avoid cache line sharing. (#3191)
* WT-2402 Pad structures to avoid cache line sharing.

We previously used compiler alignment to avoid cache line sharing, which
is problematic because we don't allocate memory with the necessary
alignment, so some behavior is undefined.  Instead, this change merely
pads structs to ensure they won't share a cache line when allocated in
an array.

* Add some #if barriers around the new union's

* Change where CACHE_LINE_PAD defines live and rename _START to _BEGIN

* Add ax_check_compile_flag to aclocal

* Lint

* Allocate the TINFO structures separately to minimize false sharing.

* Comment cleanup to make spell happy.

* Pad the read-write locks to avoid false sharing.

* Add a comment as to why we're adding -std=c11 to the compile.

* whitespace

* Add asserts that we're padding the structures we intended to pad.

Minor cleanup, don't use #defines outside of the WiredTiger namespace.