WT-2833 Add projections to wt dump utility (#3192)
When dumping with a projection, the dumped metadata associated with the
table is modified so that the list of columns matches the list from the
projection, and the value format corresponds to projected format.  We use an
open cursor using the projection to obtain the value format to show, that
requires a slight reordering of when cursors are opened in the dump utility.

Also fix a problem in the JSON dump cursor which did not handle projections.

Added tests for dump with projections, including tables and indices using
the normal dump formats, and tables dumped using JSON dump format. Tables
are also dumped with projections and then loaded into the modified table
format to be checked.