WT-3134 Coverity scan reports 1368529 and 1368528 (#3251)
* Coverity complains:
	CID 1368529:  Security best practices violations  (TOCTOU)
	Calling function "fopen" that uses "fname" after a check
	function. This can cause a time-of-check, time-of-use race
	condition.

We're doing:

	snprintf(buffer);
	stat(buffer);
	snprintf(buffer);
	fopen(buffer);

and I think Coverity is ignoring the second snprintf(), and is complaining
about a stat followed by an fopen some number of lines of code later.  It's
simple enough to give the two calls their own buffers, hopefully that will
keep Coverity quiet.

Use 1024 as the size of a path instead of 512, (that's the traditional
MAXPATHLEN value).

Use sizeof(home) in calls to testutil_work_dir_from_path() so we don't
accidentally diverge from the declared size.

Clean up an error call, there's no need for two error messages.

* Coverity complains:

	CID 1368528:    (DEADCODE)
	Execution cannot reach the expression "","" inside this
	statement: "pos += (uint32_t)snprintf(c...".

Replace boolean variable with a "const char *" that's set to either an
empty string or a comma, removing the need for the test.

Use size_t as the size of an object in memory, not a uint32_t.

Don't declare variables in block scope.

Assignment operators are the lowest priority operator (well, except for
comma), don't bother declaring the order of evalution for an assignment
operator.

strlen() returns a size_t length, don't evaluate it as a boolean.