SERVER-27807 synchronize creating a snapshot with its registration in replcoord
This commit prevents a race between creating a snapshot in the storage engine and
registering it in the replication coordinator.  The replication coordinator maintains
a vector of outstanding snapshots, and it needs to stay in sync with the actual snapshots
in the storage engine.  The replication coordinator mutex is used to ensure this synchronization.

(cherry picked from commit 8253fab192fad307a07846878e368e970990d7b3)