WT-3115 Convert the dhandle list lock into a read/write lock. (#3236)
It was a spinlock, but most acquirers only need shared access and it can be a contention point in many-table workloads.

Split uses of the handle list lock into small operations. In particular, only hold the handle list lock to get the "next" handle, not for loops over all the handles in the system.

Update statistics around handle list lock and corresponding doc.