WT-3149 Have eviction choose a random point when walking a tree. (#3285)
Only choose a random point when there is no saved walk point.

Fixes to random search as well - noticed search termination conditions when sampling the search page vs. walking it sequentially weren't the same. Changed that, which caused the
test_compact02 test to fail.

There's an underlying bug in this code, if we return WT_NOTFOUND, we can
lose a hazard pointer on the page of the tree we unsucessfully searched.
Add a page-release in the case of returning not-found.