WT-3204 eviction changes cost LSM performance (#3325)
* WT-3204 eviction changes cost LSM performance

Modify LSM's primary chunk switching to match the new btree eviction
semantics on object creation. We now create objects with eviction turned
off, LSM should no longer have to turn eviction off when configuring the
primary chunk.

LSM previously set WT_BTREE.bulk_load_ok to false to ensure an insert
into the tree wouldn't turn eviction on. That problem remains, but
there's a race in the implementation if multiple threads are inserting
at the same time (where a thread modifies WT_BTREE.bulk_load_ok and goes
to sleep before configuring eviction, and another thread does an insert
and turns off eviction), and there's a further race between threads
doing F_ISSET/F_SET tests. Change the WT_BTREE_LSM_PRIMARY flag into a
WT_BTREE.lsm_primary variable so there's no F_ISSET/F_SET race. Remove
the test/set of bulk-load_ok, instead, test the lsm_primary value in the
btree code before turning eviction off.

When checkpointing an LSM chunk, move the code that turns off the
chunk's primary flag in the chunk inside the single-threaded part of the
function to ensure we don't race with other threads doing checkpoints.
That makes the code to fix up the accounting single-threaded and safe.

Simplify the LSM checkpoint code to call __wt_checkpoint directly, and
use the same handle for turning off the chunk's primary flag as we use
for the checkpoint.

* Force a primary switch in LSM after an exclusive-handle operation
has come through.

Otherwise it's possible to attempt to use a file as the primary chunk
without disabling eviction.

* spelling

* WT_BTREE.bulk_load_ok isn't a boolean, don't use true/false comparisons.

* Only check for an empty tree the first time an LSM chunk is opened.

The goal here is to make sure that LSM primary chunks start empty.
Otherwise, we can't load into a skiplist in memory as required by LSM.
If an operation such as verify closes a btree in order to check the
on-disk state, the next time it is opened we have to check whether it is
empty.

It is safe to do this check without locking: what matters is that we
always do the `lsm_primary` check before any update operation that would
turn off `btree->bulk_load_ok`.

* Rename WT_BTREE.bulk_load_ok to be WT_BTREE.original, it's used by LSM.

* Fix a comment.