WT-3196 Prevent eviction in LSM primaries after the are flushed. (#3336)
Once an LSM primary is known to be on disk, we expect readers to use the
checkpoint.  The original page image for the primary will then be
discarded by an LSM worker thread.

We previously allowed the LSM primary to be evicted in between so that
eviction workers can deal with cache pressure ahead of the LSM worker
threads discarding the chunk.  However, that leads to cases where
application threads end up evicting a 100MB page, and also means that
discarding the chunk needs to worry about split generations (the cause
of the assertion failure here).

The solution suggested here is simple: never enable eviction in LSM
primaries, which also means we never need to fix up cache accounting.