SERVER-27761 track collation in ViewGraph
Keeps track of the collation of views in the ViewGraph, to ensure that all views in the same
connected component share the same collation. This prohibits users from dropping a view and
recreating it with a different collation if other views depend on it.

(cherry picked from commit 476c7733f5dc01011e1ebd9d7818aad8d63a1781)

 Conflicts:
	jstests/core/views/views_collation.js
	src/mongo/db/views/SConscript
	src/mongo/db/views/view_graph.cpp