WT-3265 Allow eviction of recently split pages when tree is locked. (#3372) (cherry picked from commit: 84e6ac0e67019bba22af87b99b40bb0bc0e21157)
When pages split in WiredTiger, internal pages cannot be evicted
immediately because there is a chance that a reader is still looking at
an index pointing to the page.  We check for this when considering pages
for eviction, and assert that we never evict an internal page in an
active generation.

However, if a page splits and then we try to get exclusive access to
the tree (e.g., to verify it), we could fail to evict the tree from
cache even though we have guaranteed exclusive access to it.

Relax the check on internal pages to allow eviction from trees that are
locked exclusive.