WT-3142 Add a workload generator application. (#3255)
The workload generator runs core workloads in C++, which directly calls the C API and thus minimizes the overhead of the test program. The workload setup is done in Python to allow for an expressive workload definition language.

Initial commit has a number of example workloads that match equivalent wtperf workloads.