WT-3248 Mark reconciliation a success if no skipped updates (#3441)
Only give up on update/restore if we don't use any updates.

Previously, we were tracking the size of saved updates vs the size of
uncommitted updates: renamed the variables for clarity.  Only give up on
update/restore eviction if (1) the page doesn't split and (2) all
updates are saved -- i.e., we're not applying anything.