WT-3354 Fix bugs found by Coverity. (#3451)
* WT-3354 Fix bugs found by Coverity.

* two cases where error checking for rwlocks should goto the error label for cleanup.
* LSM code not restoring isolation if a checkpoint fails part way through

* Take care with ordering an assertion after a read barrier.

We just had an assertion failure on PPC, and from inspection it looks
like read in the assertion could be scheduled before read that sees the
ticket allocated.  We have a read barrier in this path to protect
against exactly that kind of thing happening to application data, move
the assertion after it so our diagnostics are also safe.