SERVER-28774 Add RenameNode.
A $rename is implemented with two UpdateLeafNode classes: RenameNode and
ConflictPlaceholderNode. The ConflictPlaceHolder node exists only to
mark the source path of the rename in the update tree, so that it
generates a conflict if another modifier tries to act on the same
path.

Actually writing the destination path and unsetting the source path
occurs in RenameNode apply method.

The RenameNode gets placed in the update tree according to its
destination path, so that when we reach the RenameNode apply method,
pathTaken and pathToCreate are with respect to the destination path,
allowing us to create the path if we need to.