SERVER-30988 Fix misleading $jsonSchema error.
My original implementation of the $jsonSchema ban included a Status
error messaged that assumed the only reason to disallow $jsonSchema is
because of a featureCompatibilityVersion < 3.6. That assumption will
not be true, so this patch replaces that error with a generic
error. Code paths that include the $jsonSchema check because of the
featureCompatibilityVersion now rewrites this generic error with an
error that includes a link to the relevant documentation.

There is a ticket open to enable $jsonSchema in more places:
SERVER-30987.