WT-3582 Move test/format checkpoints to a dedicated thread. (#3697)
* WT-3582 Move test/format checkpoints to a dedicated thread.

* typo.

* If we fail to open a checkpoint, we're not incrementing session_op so
it falls behind tinfo->ops and we'll never open a checkpoint again.
Always update session_op if we're swapping the session/cursor. Also,
move the test for a running transaction into the same clause so we don't
test session_op twice, and test for GTE rather than equality to avoid
introducing the bug in the future.

Don't read from a checkpoint if it's a data-source, they don't support
that.

* g.checkpoint_lock structure field no longer used.

* Always do a checkpoint within the first 10 seconds, otherwise short runs
won't ever read/create checkpoints.

We can create named checkpoints during a backup, we just can't drop them.
Only test for the backup lock if we're going to drop the checkpoints.

Rotate between a few named snapshot names so we test having multiple named
snapshots in the system.

pthread_rwlock_trywrlock() and pthread_rwlock_unlock() can fail, test for
errors.