WT-3638 When restoring updates, don't discard on-page tombstones. (#3713)
* WT-3638 When restoring updates, don't discard on-page tombstones.

Reconciliation selects an update as the "on-page" version that is
correct as-of the reconciliation.  If that update is a tombstone, no
value will appear on the page.  Our API permits a modify of a deleted
record, so we can't discard a tombstone if it was chosen as the on-page
update, or a later modify update will have no starting value.

* Enhance the comment to clarify what's going on in update list
truncation.