WT-3652 Skip unnecessary lookaside reads / writes. (#3744)
Also fix reads during checkpoint so the expected read generation is
set, and checkpoint can attempt to clean up after itself.

Add a "lookaside_score" measuring the proportion of unstable updates in
cache (those required for historic reads), use that to determine when
to use the lookaside table rather than waiting for the cache to become
stuck.

Instead of discarding updates as part of restoring a page, including the original value in the update list if the on-page update is a modify.  This removes some problematic code that was inconsistent about removing updates.  Also, if we need to restore updates earlier than the on-page version, the previous code was incorrect.

Disable checkpoint skipping lookaside pages for now: always visit every page until we are tracking the correct IDs and timestamps to skip properly.

Ensure all trees with lookaside pages must stay dirty.