[SERVER-31942] Truncate JIRA issue description
When the update_test_lifecycle task sees changes in unreliable tests, it
creates a JIRA issue before committing the changes.The description of
the issue contains the tag changes and if there are many of them the
description text size can exceed the allowed limit for the field
(32,767 characters).

When that happens the task fails and the changes are not committed.

Instead we should truncate the description when its size exceeds the
allowed limit.