SERVER-32474 Fix double logging when using --log=buildlogger.
Exposes an additional logger that uses the test's root logger as the
parent for any TestCase instances constructor by the hook. The fixture
logger still continues to be the parent of the hook logger.

Also changes the CleanEveryN hook to explicitly get marked as a failure
if a ServerFailure exception is thrown, rather than relying on
TestReport.combine() to do so implicitly.

(cherry picked from commit 2c451638c2456a3ebc6de87a3d2fcfecbb8d92e3)