SERVER-28830 check for closed conn in cursorHandleFromId
The old implementation of DBCommandCursor used when running the shell
with --readMode compatibility calls getConnection(which checks if the
connection has been closed) to construct a DBClientCursor which it
passes to setCursor. The new variant used with --readMode command calls
setCursorHandle with the cursor ID. Check that we aren't closed in the
new variant.