SERVER-35317: Preserve minVisible values when refreshing the catalog.
Restarting the catalog throws away the "minVisibleSnapshot" info on all
collections. This is typically safe for rollback, the value is reinstantiated
when replication recovery rolls forward from the stable timestamp. However, it's
possible to read behind the stable timestamp. Currently only test commands can
do so, but the future may formally expose this case particularly for sharded
cluster snapshot reads.