SERVER-35200 Speed up steady state oplog fetching failure detection
This patch attempts to improve how quickly a secondary node in
steady state replication detects that its current sync source has failed
or become partitioned and tries to select a new sync souce. The speed of
this process can significantly impact how long it takes for a new
primary to begin committing majority writes after a previous primary
has failed or become partitioned from the replica set. This patch
improves on the old behavior by reducing the number of fetcher restarts
to 1, and also reducing the oplog 'find' request timeout used on a restart.