SERVER-37364 Temporarily disable coordinator returning decision as soon as decision is durable
Instead, the coordinator will wait for all acks before returning the decision. Re-enable when all tests have been updated to handle a decision being returned early, possibly by making a request with a higher transaction number block if it is received by a participant still in prepare for an earlier transaction.