SERVER-36865 Improved kill_rooted_or.js and kill_aggregation.js
In addition to the CannotImplicitlyCreateCollection problem described
in the Jira ticket, these tests were querying an empty collection. The
function that drops the collection was re-creating its indexes but was
not repopulating the actual documents.

This change also fixes a couple of places where exceptions were
getting squelched. Instead, the test now checks for the specific
errors that we expect to see and fails on other errors.

These changes should make the tests more sensitive to future bugs.