SERVER-38175 invariant on markKill w extra info
The OperationContext class accepts an ErrorCodes::Error for markKilled()
and deadline-setting methods. This error code may be converted to a
Status object on checkForInterruptNoAssert() or
waitForConditionOrInterruptNoAssertUntil() call. When the error code is
one that is associated with an ErrorExtraInfo object, Status
construction will fail as we expect Status construction to either
include a subclass of ErrorExtraInfo or a BSONObj from which it can be
serialized.

Until a use case comes up, add an invariant to prevent passing an error
code with extra info