SERVER-40165 set wtimeout deadlines with the precise clock
In ReplicationCoordinatorImpl::_awaitReplication_inlock, we generate our
deadline for waiting on wtimeout with the fast clock. Because we
generate the deadline with the fast clock, but actually wait on it with
the precise clock, it is possible for us to wtimeout early if there is
significant drift between the precise and fast clocks.

This is unlikely to happen very often in practice, but could happen if
the thread responsible for pushing ahead the fast clock get's
descheduled for a substantial period of time. In that case, it is
possible for the fast clock to drift multiple seconds behind the precise
clock, and to immediately timeout in
waitForConditionOrInterruptNoAssertUntil.