SERVER-39383 Add a test for speculative majority change stream secondary reads during batch application
This commit adds an integration test to verify that speculative majority change stream reads do not return incorrect results when reading concurrently with secondary batch application. The goal is to ensure that, due to the changes from SERVER-39356, these reads will read from the most recent lastApplied timestamp on secondaries.