SERVER-39492 Allow array of BSON types in encrypt.bsonType specification
Changed the IDL definition of EncryptionInfo.bsonType to use a custom type based on
MatcherTypeSet. Since it is a pretty generic type we could move it to
basic_types.idl.
Changed InternalSchemaBinDataEncryptedTypeExpression to use
MatcherTypeSet as well, rather than a single type. Parsing logic of
MatcherTypeSet allowed for a concise code.