SERVER-40335 Add waitForStepDownOnNonCommandShutdown parameter
When ReplSetTest.stopSet() kills the primary with SIGTERM, it must wait for the
primary to finish election handoff or, in cases where there is no candidate,
for the 10-second election handoff timeout to expire. Add a parameter and use
it in ReplSetTest.stopSet() to end tests quickly.

Disable waitForStepDownOnNonCommandShutdown for replica sets started directly
by resmoke.py so those sets shutdown quickly, too.