SERVER-36865 Improved kill_rooted_or.js and kill_aggregation.js
Backports the original version of this commit (on branch v4.2), so
that we don't get CannotImplicitlyCreateCollection errors when testing
v4.0.

Unlike in the original commit, this backport does not add a function
to repopulate the test collection after dropping it. The v4.0 version
of this test does not populate the collection in the first place. We
would have to backport that functionality as well if we wanted to
backport the populateCollection function.

(cherry picked from commit 369a2e0432c27b20b106056e2f2e0849cab617f0)