SERVER-43816 Push $text and $meta:"textScore" sort into the PlanStage layer.
This change involves unifying the behavior of find and
agg for validation of $meta:"textScore". In particular, find
operations no longer require a "textScore" $meta projection
in order to specify a "textScore" $meta sort. This brings
find into alignment with agg, which never had such a
restriction. It is also now legal for a find command to
sort on the field overridden by a $meta:"textScore" projection
without specifying the $meta operator in the sort pattern.

In addition:
 - Tightens validation around uses of "textScore" $meta
 projections for queries that do not specify a $text
 predicate.
 - Fixes a bug in which $natural sort/hint were not
 correctly validated.