MSI installer produces incorrect config file
If the user selects a shorter path of the log file or data directory, the written buffer is shorter than the original content, and the remaining part of the file is not truncate automatically. This leads to an incorrect config file, and thus the mongod service cannot start. 

(I hope that I haven't done some silly mistake - C++ is really not my language)